requirements.txt
affine==2.4.0
altair==6.0.0
annotated-doc==0.0.4
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.3.0
attrs==26.1.0
babel==2.18.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cligj==0.7.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.7
cycler==0.12.1
cyclopts==4.10.2
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
docstring_parser==0.18.0
docutils==0.22.4
earthengine-api==1.7.22
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.28.0
fonttools==4.62.1
fqdn==1.5.1
fsspec==2026.3.0
geojson==3.2.0
geomet==1.1.0
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-cloud-core==2.5.1
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
h11==0.16.0
hf-xet==1.4.3
html2text==2025.4.15
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.11
ImageIO==2.37.3
ipykernel==7.2.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
json5==0.14.0
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.5.0
lark==1.3.1
lazy-loader==0.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
mpmath==1.3.0
narwhals==2.19.0
nbclient==0.10.4
nbconvert==7.17.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
notebook==7.5.5
notebook_shim==0.2.4
numpy==2.4.4
opencv-python==4.13.0.92
packaging==26.1
pandas==3.0.2
pandocfilters==1.5.1
parso==0.8.6
pillow==12.2.0
platformdirs==4.9.6
plotly==6.7.0
pooch==1.9.0
prometheus_client==0.25.0
prompt_toolkit==3.0.52
proto-plus==1.27.2
protobuf==7.34.1
psutil==7.2.2
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydeck==0.9.2
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.1.0
pyvista==0.47.3
pywinpty==3.0.3
PyYAML==6.0.3
pyzmq==27.1.0
rasterio==1.5.0
referencing==0.37.0
requests==2.33.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==15.0.0
rich-rst==1.3.2
rpds-py==0.30.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
scooby==0.11.0
Send2Trash==2.1.0
sentinelsat==1.2.1
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
soupsieve==2.8.3
stack-data==0.6.3
streamlit==1.56.0
sympy==1.14.0
tenacity==9.1.4
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2026.4.11
tinycss2==1.4.0
toml==0.10.2
torch==2.11.0
torchvision==0.26.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.2
typing_extensions==4.15.0
tzdata==2026.1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.6.3
vtk==9.6.1
watchdog==6.0.0
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
widgetsnbextension==4.0.15
